requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
xlsxwriter==3.1.9
openai==1.12.0
python-dotenv==1.0.1
